Create a Multi-Master setup document.
The doc mainly describes very simple naive setups where the
servers run on different host:port combinations but
share common DB and git repo back-ends.  Every setup
currently described in this doc does not actually require
any plugin support.  This doc is meant as a starting point
for multi-master setups and as features are added to the
multi-master plugin, this doc will be enhanced to describe
how these setups can be improved.

Change-Id: I1011f553db3cc984dbf2e9654af898edbbc840d0