Create a DEGRADED mode
This enables running a master in a DEGRADED mode, in which certain
caches are essentially disabled (or seriously degraded) by being
periodically flushed so that peer coherency issues simply disappear
at the cost of efficiency.
To enable DEGRADED mode, set the rate at which the caches should be
flushed using "cache.flushRate" in 'etc/multimaster.config'.

Change-Id: If494c0c2181213496b0414b8681879eb9aa9334c