Rename config.md to about.md
This change prepare the grounds for merging 4 plugins into this one:
evict-cache, sync-events sync-index and websession-flatfile.

Those plugins, used together, allows to run a multi-master in failover
topology. This topology is first one that this multi-master plugin will
support.

Each of these plugins have a config.md which document different
configuration options. The config.md in this plugin was more about
general information how to setup a multi-master so move it as about.md
to clear the space to document the merged plugins configuration options.

Once 4 plugins are merged, config.md and about.md will be reworked to
document how this plugins can be used in the general context of
multi-master.

Change-Id: Ia2c4860c9ac38477537525ee46d7812bfd26a1c1