Update patch set 4
Patch Set 4:

> > Another reason is that individually, each plugin have no real use
 > if not combined with the others. Also the plugins have a lot of
 > code duplication and I did not want to come up with a library to
 > fix the code duplication, it's too much work compared to have a big
 > plugin(not really big in fact, about 1000 lines of code for 2
 > plugins merged so far)
 > 
 > Got the point and makes sense. The name "multi-master", however,
 > doesn't really match the description of what the plugin does. Shall
 > we just rename it?

One of the other intentions was to have the active-passive solution as an interim step towards multimaster.

Whether or not multimaster can actually be realised by a plugin in future is another question, of course.

With that in mind maybe it is a good idea to have the plugin named "high-availability" or something like that.

Patch-set: 4
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>