Update patch set 4
Patch Set 4:

> > With that in mind maybe it is a good idea to have the plugin
 > named
 > > "high-availability" or something like that.
 > 
 > Or, just let all as is and add this current known limitation to the
 > plugin description.

I considered doing this, calling it "high-availability" but when I was
looking for the existing plugins to make sure my new plugin name would
not clash with an existing one, I stumbled on the already existing
multi-master plugin.

The plugin was empty with only this file documenting how to setup multiple
gerrit masters sharing database and repos:
https://gerrit.googlesource.com/plugins/multi-master/+/refs/heads/master/src/main/resources/Documentation/about.md

That is why I decided to merge the "failover" plugins into the multi-master.

Maybe that documentation should not have been in that plugin in the first place.

My idea was to make the multi-master plugin support different modes:

-failover cluster: 2 gerrit servers sharing database and repos, one active and one warm stand-by
-load balancing cluster: n gerrit servers sharing database and repos, all active
-Multiple clusters: each cluster have their own database and repos.

That said, if there is a strong opposition to use the multi-master plugin for that,
I can rename it to something else right away.

Another option is to keep the multi-master name and if someone comes with a multi-master
implementation(3rd mode mentioned above) that has nothing in common with the multi-master
plugin, we could rename it only when that day happens.

Patch-set: 4
Reviewer: Gerrit User 1012987 <1012987@173816e5-2b9a-37c3-8a2e-48639d4f1153>