Update patch set 4
Patch Set 4:

> I considered doing this, calling it "high-availability" but when I was
> looking for the existing plugins to make sure my new plugin name would
> not clash with an existing one, I stumbled on the already existing
> multi-master plugin.

Ah, now I understand. I believe the multi-master plugin was created during a Hackathon years ago, when some companies said they were willing to invest into the project ... but then was left almost empty.

> 
> The plugin was empty with only this file documenting how to setup multiple
> gerrit masters sharing database and repos:
> https://gerrit.googlesource.com/plugins/multi-master/+/refs/heads/master/src/main/resources/Documentation/about.md
> 
> That is why I decided to merge the "failover" plugins into the multi-master.
> 
> Maybe that documentation should not have been in that plugin in the first place.
> 
> My idea was to make the multi-master plugin support different modes:
> 
> -failover cluster: 2 gerrit servers sharing database and repos, one active and one warm stand-by

yes, this is generally referred as High-Availability with Active/Passive failover.

> -load balancing cluster: n gerrit servers sharing database and repos, all active

yes, this is generally referred as WLB configuration and provides the out-of-the-box high availability. They are all the "same" master cluster because the master repository is only one.

> -Multiple clusters: each cluster have their own database and repos.

yes, and this is what is typically referred as "multi-master" in most of the Gerrit discussions. 

I believe the 1st and 2nd could be called "load-balancing" scheme whilst only the last is a multi-master.

> That said, if there is a strong opposition to use the multi-master plugin for that,
> I can rename it to something else right away.

Not strong opposition, just possibly make sure that the plugin description says that this is a load-balancing support for Gerrit and *NOT* a multi-master.

> 
> Another option is to keep the multi-master name and if someone comes with a multi-master
> implementation(3rd mode mentioned above) that has nothing in common with the multi-master
> plugin, we could rename it only when that day happens.

We started working on a "pluggable" repository implementation at the last Hackathon in Mountain View: you would need to replace the repository manager and that would need a libmodule rather than a plugin.

I doubt we will be able to provide a multi-master setup in Gerrit with a simple plugin, but I may be wrong :-)

Patch-set: 4
Reviewer: Luca Milanesio <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>