Update patch set 4
Patch Set 4:

My main goal is to merge the failover plugins into one plugin, this will support
failover mode and short term, I will also work to support co-located active
servers. I will not work on the multi-master mode and as you guys said, it's
unlikely that multi-master will be implemented in a plugin.

I want to move forward with merging the 2 other plugins but we need to agree on
the plugin name.

2 options
-keep multi-master
-create a new plugin "high-availability", move the changes done so far and revert
the multi-master plugin to the way it was before I started to merge my plugins into it.

After reading all the arguments, I like "high-availability", it's more clear what
the plugin is for and the name make sense for the 2 modes I want to support.

Let's hear what you guys vote for.

Patch-set: 4
Reviewer: Hugo Arès <1012987@173816e5-2b9a-37c3-8a2e-48639d4f1153>